refactor(explorer): split object explorer protocol from picker UI
Extracts low-level LSP Object Explorer session, cache management, and
mapping variables out of `find_object.lua` and moves them into a new
`explorer.lua` module.

Updates `lsp.lua`, `query_manager.lua` and `autocmds.lua` to import the
new `explorer` module and direct schema caching, expansion
notifications, and buffer-level cache pruning calls to the protocol
layer.
